Major hack to deal with resampling of uncollapsed parameters.
Even though we will resample the parameters we need to compute the
probability under the current parameter assignemnt, and then
transition them. This behavior is correct if you think about the
next time that the Gibbs kernel arrives at this location, it will
compute the logp of the resampled parameters last time, then resmaple
as desired.