Rewrite lognormal in terms of normal, eliminating code reuse.
Lognormal transfers its incorporates into the log domain and
invokes normal methods. Care should be taken for the log(x) and
sum_log_x prefactors in the density evaluation. The simulation is
yet to be tested. Could consider making lognormal a subclass of
normal, but their attributes (sufficient statistics) would be named
differently, so currently believe that is not the righ approach.