Eliminate lower quality uncollapsed samplers.
The code can be restored if necessary. Ultimately the cost of
maintaining and working around these two uncollapsed samplers was
get too high as I am designing the SMC algorithm, and its hard to
imagine a practical use case where one would use the uncollapsed
exponential/normal.

The remaining issue is the BetaUC which does not have a conjugate
model but feels easier to manage at least.