Add reassign flag for dim_incorporate.
The previous state of the world was that incorporate_dim would
always invoke reassign on the dim, which resamples any uncollapsed
parameters. This behavior violated the correctness of the column
kernel for uncollapsed dimensions, since we effectively were replacing
the parameters that were sampled for the partition of each view.