Implement INCORPORATE ROW semantics.
A big change is coming soon with respect to this notion of a global
data store. In particular, if a collection of GPMs are to be composed
all acting on the same dataset, we need the global variable to be
shared properly. Currently there is a hacky distinction between operating
on existing rowids, and creating new rowids. If all the GPMs agree
to operate on the same data store, then this distinction will immediately
go away and we can have all of dim, view, and state be the GPMs on slices
of that data store.