Do not maintain pointers to Dims in the State GPM.
The rationale: when calling view.incorporate_dim(D), the view should
be free to make whatever modifications it needs to D, including creating
a clone of D that is different than the object passed in by the State.
By avoiding two pointers to D (in state.dims and view.dims) we can avoid
any inconsistency issues associated with reassigning.

An alternative would be passing in a pointer (ie a Dim in a list) which
I just thought of while writing this commit message.