Fix test_simulate_mixture by avoiding Engine and seeding the State.
There appears to be an issue picking up inference from an Engine whose
random seed is reset. This will largely be resovled when we start saving
the rng object itself rather than tampering with global state.

Interestingly in cpp-crosscat, transition-stop-transition has been shown
to cause huge jumps in the logscore, and Baxter never figured out why that
was the case.