transition_hyper_grids should only transition_hypers if uninitialized.
This fix addresses the observation that engine.get_state was clobbering the
hypers. Since we are calling transition_hyper_grids in the init of State, we end
up ignoring the serialized hypers, and instead resample them uniformly from the
grid in Dim. It also explains why calling state.transition_column_hypers was
recalibrating the column hypers after loading them.

test_simulate_mixture now uses Engine and get_state to ensure that the hypers
are correctly loaded and not resampled.