Add tests for restricted CRP fix bugs in existing implementation.
One hack is that re-using code for the column CRP to the row CRP will
require specifying the number of rows, which is not too complex from
an interface level. The current solution is to allow the user to enter
None for N and set the number of rows to 1e9. When the views use the
constrained CRP to generate their row partitions, their Rd/Ri is going
to be empty dictionaries anyway (no nesting).