Fix RNG in DistributionGpm and wrap noise model around random forest.
- All the DistributionGpm objects now accept an RNG on initialization.
- Rewrote random forest to deal with degenerate model much better by
having a stochastic outelier variable \alpha upon which we do inference.
The metaphor of hyperparams, params, and data is now fullfilled by the
RandomForest where transition_params just does ML estimation of the random
forest, and the transition_hypers goes through the grid of \alpha (uniform)
to sample the outlier probability.