Fix security vulnerability. Our collaborator identified the of sqlalchemy as problematic.
Conda says, when building the notebook image, that bayeslite has an indirect dependency
to sqlalchemy. However, when I make this dependency explicit, bayeslite complains that
CGPM doesn't like. Thus pushing to a branch here to see the conda  build
on travis.

Sosqlalchemy making this dependency explicit and bumping it's version should solve
the problem.