Removing `items()` as in Python3 this will return a view object that is not JSON serializable.
fsaad suggested to do do tuple(dict.items()) or list(dict.items()) to convert
the view/iterator object into an explicit tuple or list.

I am in two minds to do this here. My sense is that the conversion should be done by
the client (i.e. convert_cgpm.py and the function calling/feeding it). I
rather have the output of state.py and crp.py faithful/type checking
with the data structures used in the CGPM modules.